refactor weather page
