+++ src/pages/WeatherPage.css
@@ -19,7 +19,6 @@
 
 .weather-description {
   margin-top: 10px;
-  text-transform: capitalize;
   font-size: 20px;
 }
 
@@ -31,8 +30,8 @@
   display: inline-flex;
   justify-content: center;
   align-items: center;
-  height: 90px;
-  width: 90px;
+  height: 95px;
+  width: 95px;
   border-radius: 50%;
   color: #fff;
   background-color: #2c79ae;
