+++ src/utils/utils.ts
@@ -1,0 +1,3 @@
+export function capitalize(value: string): string {
+  return value[0].toUpperCase() + value.slice(1);
+}
